chore: pull out errors
