+++ context/errors.go
@@ -1,4 +1,4 @@
-package internal
+package context
 
 import "errors"
